requirements.txt
bs4
xmltodict
openai
langchain >=0.2.0
langchain-community >=0.2.0
langchain-openai >=0.1.0
faiss-cpu >=1.7.4
tiktoken
requests
beautifulsoup4
